src/requirements.txt
numpy==1.26.4
Pillow==10.0.0
urllib3==2.3.0
wandb==0.19.7
albumentations==2.0.5
seaborn==0.13.2
boto3==1.37.0
sagemaker==2.241.0